feat: generate wrapper script and temp script to update GOROOT and PATH

+++ main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
-	"golang.org/x/sys/windows/registry"
 	"github.com/danielkermode/gvm/web"
+	"golang.org/x/sys/windows/registry"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -12,16 +12,61 @@ import (
 )
 
 const (
-	GvmVersion = "1.0.0"
+	GvmVersion = "2.0.0"
 )
 
-//callback function for looping over files. If true, breaks the loop.
+var (
+	GvmScript     = "gvm"
+	GvmTempScript = "gvm$"
+)
+
+// callback function for looping over files. If true, breaks the loop.
 type callback func(file os.FileInfo, reg *regexp.Regexp, proot string) bool
 
 func main() {
 	args := os.Args
 	osArch := strings.ToLower(os.Getenv("PROCESSOR_ARCHITECTURE"))
 	detail := ""
+
+	//warbaby
+	executable, _ := os.Executable()
+	exe := ""
+	if strings.HasSuffix(executable, ".exe") {
+		exe = ".exe"
+		GvmScript += ".bat"
+		GvmTempScript += ".bat"
+	}
+	dir := filepath.Dir(executable)
+	innerGoName := filepath.Join(dir, "gvm0"+exe)
+	batchName := filepath.Join(dir, GvmScript)
+
+	base := filepath.Base(executable)
+
+	if base == "gvm" || base == "gvm.exe" {
+		err := os.Remove(innerGoName)
+		if err == nil || os.IsNotExist(err) {
+			err = os.Rename(executable, innerGoName)
+		}
+		if err != nil {
+			fmt.Printf("Cannot rename %s to %s: %s", executable, innerGoName, err.Error())
+			os.Exit(-1)
+			return
+		}
+		s := "@echo running gvm0 from wrapper script...\n" +
+			"@gvm0 %*\n" +
+			"@IF EXIST \"gvm$.bat\" (\n" +
+			"    @call gvm$.bat\n" +
+			"    @del gvm$.bat\n" +
+			")"
+		err = os.WriteFile(batchName, []byte(s), 755)
+		if err != nil {
+			fmt.Printf("Cannot write file %s: %s", batchName, err.Error())
+			os.Exit(-1)
+			return
+		}
+		fmt.Println("Running gvm not gvm0, rename to gvm0 and generate wrapper script...\nDone, please run again!")
+		return
+	}
 
 	if osArch == "x86" {
 		osArch = "386"
@@ -48,6 +93,8 @@ func main() {
 			fmt.Println("Successfully installed Go version " + detail + ".")
 			fmt.Println("To use this version, run gvm use " + detail + ". This will also set your GOROOT.")
 		}
+	case "gvmroot":
+		gvmroot(detail)
 	case "goroot":
 		goroot(detail)
 	case "list":
@@ -69,16 +116,32 @@ func main() {
 
 func install(version string, arch string) bool {
 	fmt.Println("")
-	if os.Getenv("GOROOT") == "" {
-		fmt.Println("No GOROOT set. Set a GOROOT for Go installations with gvm goroot <path>.")
+
+	if os.Getenv("GVMROOT") == "" {
+		fmt.Println("No GVMROOT set. Set a GVMROOT for go installations with gvm gvmroot <path>.")
 		return false
 	}
 	if version == "" {
 		fmt.Println("Version not specified.")
 		return false
 	}
-	gorootroot := filepath.Clean(os.Getenv("GOROOT") + "\\..")
-	return web.Download(version, "windows-"+arch, gorootroot)
+
+	return web.Download(version, "windows-"+arch, root())
+}
+
+func gvmroot(path string) {
+	fmt.Println("")
+	if path == "" {
+		if root() == "" {
+			fmt.Println("No GVMROOT set.")
+		} else {
+			fmt.Println("GVMROOT: ", root())
+		}
+		return
+	}
+	path = filepath.FromSlash(path)
+	_ = os.WriteFile(GvmTempScript, []byte(fmt.Sprintf("SET GVMROOT=%s\nSETX GVMROOT %s\n", path, path)), 0755)
+	fmt.Println("Set the GVMROOT to " + path + ".")
 }
 
 func goroot(path string) {
@@ -88,36 +151,40 @@ func goroot(path string) {
 			fmt.Println("No GOROOT set.")
 		} else {
 			fmt.Println("GOROOT: ", os.Getenv("GOROOT"))
-			fmt.Println("Other Go versions installed at: ", filepath.Clean(os.Getenv("GOROOT")+"\\.."))
 		}
 		return
 	}
-	newpath := filepath.FromSlash(path)
-	//permanently set env var for user and local machine
-	//The path should be the same for all windows OSes.
-	machineEnvPath := "SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment"
-	userEnvPath := "Environment"
-	setEnvVar("GOROOT", newpath, machineEnvPath, true)
-	setEnvVar("GOROOT", newpath, userEnvPath, false)
-	//Also update path for user and local machine
-	updatePathVar("PATH", filepath.FromSlash(os.Getenv("GOROOT")), newpath, machineEnvPath, true)
-	updatePathVar("PATH", filepath.FromSlash(os.Getenv("GOROOT")), newpath, userEnvPath, false)
-	fmt.Println("Set the GOROOT to " + newpath + ". Also updated PATH.")
-	fmt.Println("Note: You'll have to start another prompt to see the changes.")
+	newRoot := filepath.FromSlash(path)
+	newBin := newRoot + string(filepath.Separator) + "bin"
+	oldBin := os.Getenv("GOROOT") + string(os.PathSeparator) + "bin"
+
+	pathEnv := updatePath(oldBin, newBin)
+	_ = os.WriteFile(GvmTempScript, []byte(fmt.Sprintf("SET GOROOT=%s\nSET PATH=%s\n", newRoot, pathEnv)), 755)
+	fmt.Println("Set the GOROOT to " + newRoot + ". Also updated PATH.")
+}
+
+func root() string {
+	return filepath.Clean(os.Getenv("GVMROOT"))
 }
 
 func listGos() {
-	if os.Getenv("GOROOT") == "" {
-		fmt.Println("No GOROOT set. Set a GOROOT for go installations with gvm goroot <path>.")
+	if root() == "" {
+		fmt.Println("No GVMROOT set. Set a GVMROOT for go installations with gvm gvmroot <path>.")
 		return
 	}
+
+	fmt.Printf("listing go versions in '%s'\n", root())
+
 	//store all Go versions so we don't list duplicates
 	goVers := make(map[string]bool)
 
-	callb := func(f os.FileInfo, validDir *regexp.Regexp, gorootroot string) bool {
+	callb := func(f os.FileInfo, validDir *regexp.Regexp, gvmroot string) bool {
 		if f.IsDir() && validDir.MatchString(f.Name()) {
-			goDir := filepath.Join(gorootroot, f.Name())
+			goDir := filepath.Join(gvmroot, f.Name())
 			version := getDirVersion(goDir)
+			if version == "" {
+				return false
+			}
 			//check if the version already exists (different named dirs with the same go version can exist)
 			_, exists := goVers[version]
 			if exists {
@@ -125,9 +192,12 @@ func listGos() {
 			}
 			str := ""
 			if goDir == os.Getenv("GOROOT") {
-				str = str + "  * " + version[2:] + " (Using with GOROOT " + goDir + ")"
+				str += "  * " + version[2:]
 			} else {
-				str = str + "    " + version[2:]
+				str += "    " + version[2:]
+			}
+			if version != f.Name() {
+				str += " \t in " + f.Name()
 			}
 			goVers[version] = true
 			fmt.Println(str)
@@ -139,19 +209,23 @@ func listGos() {
 }
 
 func uninstall(unVer string) {
+	if root() == "" {
+		fmt.Println("No GVMROOT set. Set a GVMROOT for go installations with gvm gvmroot <path>.")
+		return
+	}
 	if unVer == "" {
 		fmt.Println("A version to uninstall must be specified.")
 		return
 	}
 
-	callb := func(f os.FileInfo, validDir *regexp.Regexp, gorootroot string) bool {
+	callb := func(f os.FileInfo, validDir *regexp.Regexp, gvmroot string) bool {
 		if f.IsDir() && validDir.MatchString(f.Name()) {
-			goDir := filepath.Join(gorootroot, f.Name())
+			goDir := filepath.Join(gvmroot, f.Name())
 			version := getDirVersion(goDir)
 			if version == "go"+unVer {
 				os.RemoveAll(goDir)
 				fmt.Println("Uninstalled Go version " + version[2:] + ".")
-				fmt.Println("Note: If this was your GOROOT, make sure to set a new GOROOT with gvm goroot <path>")
+				fmt.Println("Note: If this was your GOROOT, you should use other go version with gvm use <version>")
 				return true
 			}
 		}
@@ -165,30 +239,28 @@ func uninstall(unVer string) {
 }
 
 func useGo(newVer string) {
-	if os.Getenv("GOROOT") == "" {
-		fmt.Println("No GOROOT set. Set a GOROOT for go installations with gvm goroot <path>.")
+	if root() == "" {
+		fmt.Println("No GVMROOT set. Set a GVMROOT for go installations with gvm gvmroot <path>.")
 		return
 	}
 	if newVer == "" {
 		fmt.Println("A new version must be specified.")
 		return
 	}
-	callb := func(f os.FileInfo, validDir *regexp.Regexp, gorootroot string) bool {
+	callb := func(f os.FileInfo, validDir *regexp.Regexp, gvmroot string) bool {
 		if f.IsDir() && validDir.MatchString(f.Name()) {
-			goDir := filepath.Join(gorootroot, f.Name())
+			goDir := filepath.Join(gvmroot, f.Name())
 			version := getDirVersion(goDir)
 			if version == "go"+newVer {
-				//permanently set env var for user and local machine
-				//The path should be the same for all windows OSes.
-				machineEnvPath := "SYSTEM\\CurrentControlSet\\Control\\Session Manager\\Environment"
-				userEnvPath := "Environment"
-				setEnvVar("GOROOT", filepath.FromSlash(goDir), machineEnvPath, true)
-				setEnvVar("GOROOT", filepath.FromSlash(goDir), userEnvPath, false)
-				//Also update path for user and local machine
-				updatePathVar("PATH", filepath.FromSlash(os.Getenv("GOROOT")), goDir, machineEnvPath, true)
-				updatePathVar("PATH", filepath.FromSlash(os.Getenv("GOROOT")), goDir, userEnvPath, false)
+				newRoot := filepath.FromSlash(goDir)
+				newBin := newRoot + string(filepath.Separator) + "bin"
+				oldBin := os.Getenv("GOROOT") + string(os.PathSeparator) + "bin"
+
+				pathEnv := updatePath(oldBin, newBin)
+				_ = os.WriteFile(GvmTempScript, []byte(fmt.Sprintf("SET GOROOT=%s\nSET PATH=%s\n", newRoot, pathEnv)), 755)
+
 				fmt.Println("Now using Go version " + version[2:] + ". Set GOROOT to " + goDir + ". Also updated PATH.")
-				fmt.Println("Note: You'll have to start another prompt to see the changes.")
+
 				return true
 			}
 		}
@@ -202,16 +274,29 @@ func useGo(newVer string) {
 
 func loopFiles(fn callback) bool {
 	validDir := regexp.MustCompile(`go(\d\.\d\.\d){0,1}`)
-	gorootroot := filepath.Clean(os.Getenv("GOROOT") + "\\..")
-	files, _ := ioutil.ReadDir(gorootroot)
+	files, _ := ioutil.ReadDir(root())
 	fmt.Println("")
 	for _, f := range files {
-		shouldBreak := fn(f, validDir, gorootroot)
+		shouldBreak := fn(f, validDir, root())
 		if shouldBreak {
 			return true
 		}
 	}
 	return false
+}
+
+func updatePath(oldBin, newBin string) string {
+	oldInfo, _ := os.Stat(oldBin)
+	path := os.Getenv("PATH")
+	pvars := strings.Split(path, ";")
+	nvars := make([]string, 0, len(pvars))
+	for _, pvar := range pvars {
+		check, _ := os.Stat(pvar + string(os.PathSeparator) + "bin")
+		if !os.SameFile(oldInfo, check) {
+			nvars = append(nvars, pvar)
+		}
+	}
+	return newBin + string(os.PathListSeparator) + strings.Join(nvars, ";")
 }
 
 func setEnvVar(envVar string, newVal string, envPath string, machine bool) {
@@ -272,14 +357,19 @@ func updatePathVar(envVar string, oldVal string, newVal string, envPath string, 
 }
 
 func getDirVersion(dir string) string {
-	files, _ := ioutil.ReadDir(dir)
+	files, _ := os.ReadDir(dir)
 	for _, f := range files {
 		if f.Name() == "VERSION" {
-			dat, err := ioutil.ReadFile(filepath.Join(dir, f.Name()))
+			dat, err := os.ReadFile(filepath.Join(dir, f.Name()))
 			if err != nil {
 				return "Error reading file."
 			}
-			return string(dat)
+			split := strings.Split(string(dat), "\n")
+			if len(split) >= 0 {
+				return split[0]
+			} else {
+				return ""
+			}
 		}
 	}
 	return ""
@@ -291,6 +381,7 @@ func help() {
 	fmt.Println(" ")
 	fmt.Println("  gvm arch                     : Show architecture of OS.")
 	fmt.Println("  gvm install <version>        : The version must be a version of Go.")
+	fmt.Println("  gvm gvmroot [path]           : Sets GVMROOT environment variable, also update windows Registry.")
 	fmt.Println("  gvm goroot [path]            : Sets/appends GOROOT/PATH. Without the extra arg just shows current GOROOT.")
 	fmt.Println("  gvm list                     : List the Go installations at or adjacent to GOROOT. Aliased as ls.")
 	fmt.Println("  gvm uninstall <version>      : Uninstall specified version of Go. If it was your GOROOT/PATH, make sure to set a new one after.")
